+++ Car/android/app/build.gradle
@@ -130,7 +130,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.softservecarmobile"
+        applicationId "com.car"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
